fix(SSR component cache): cache the rendered component as a string, stop cached child components reg
The component cache only store the rendered component as a string. When we take back the component
from the cache, we do not register its child components to their lifecycle callbacks, they are
cached, so they are not dynamic anymore in the server. When the browser handle the view, they become
dynamic, as expected.

fix #7595